@@ -1,11 +1,12 @@
-import {Transaction} from  "algosdk/src/transaction";
+import { Transaction } from 'algosdk';
+import AnyTransaction from 'algosdk/dist/types/src/types/transactions';
 
-export function buildTransaction(txn: object) {
-    let builtTxn = new Transaction(txn);
-    if(txn["group"]) {
-        // Remap group field lost from cast 
-        builtTxn.group = Buffer.from(txn["group"],'base64');
-    }
+export function buildTransaction(txn: object): Transaction {
+  const builtTxn = new Transaction(txn as AnyTransaction);
+  if (txn['group']) {
+    // Remap group field lost from cast
+    builtTxn.group = Buffer.from(txn['group'], 'base64');
+  }
 
-    return builtTxn;
+  return builtTxn;
 }
